# Add SIMD-optimized "is this well-formed UTF-8?" logic